ci-operator multi-stage: move to separate package
This has become a world of its own (from about 200 to 1000 lines and
growing).  Move to its own package so it can be separated into several
files.
